SW-5091 Fix the vertical sizing of the phase score header when there isn't a 'modifiedTime'
